airflow: expose parent run information as macros in Airflow 2
