+++ integration/airflow/README.md
@@ -192,10 +192,13 @@ job runs and dataset inputs/outputs. To propagate the job hierarchy, tasks must 
 id so that the downstream process can report the [ParentRunFacet](https://github.com/OpenLineage/OpenLineage/blob/main/spec/OpenLineage.json#/definitions/ParentRunFacet)
 with the proper run id.
 
-The `lineage_run_id` macro exists to inject the run id of a given task into the arguments sent to a
-remote processing job's Airflow operator. The macro requires the DAG run_id and the task to access
-the generated run id for that task. For example, a Spark job can be triggered using the
-`DataProcPySparkOperator` with the correct parent run id using the following configuration:
+The `lineage_run_id` and `lineage_parent_id` macros exists to inject the run id or whole parent run information
+of a given task into the arguments sent to a  remote processing job's Airflow operator. The macro requires the 
+DAG run_id and the task to access the generated run id for that task. For example, a Spark job can be triggered
+using the `DataProcPySparkOperator` with the correct parent run id using the following configuration:
+
+Airflow 1.10:
+
 ```python
 t1 = DataProcPySparkOperator(
     task_id=job_name,
@@ -206,6 +209,20 @@ t1 = DataProcPySparkOperator(
             f"-javaagent:{jar}={os.environ.get('OPENLINEAGE_URL')}/api/v1/namespaces/{os.getenv('OPENLINEAGE_NAMESPACE', 'default')}/jobs/{job_name}/runs/{{{{lineage_run_id(run_id, task)}}}}?api_key={os.environ.get('OPENLINEAGE_API_KEY')}"
         dag=dag)
 ```
+
+Airflow 2.0+:
+
+```python
+t1 = DataProcPySparkOperator(
+    task_id=job_name,
+    #required pyspark configuration,
+    job_name=job_name,
+    dataproc_pyspark_properties={
+        'spark.driver.extraJavaOptions':
+            f"-javaagent:{jar}={os.environ.get('OPENLINEAGE_URL')}/api/v1/namespaces/{os.getenv('OPENLINEAGE_NAMESPACE', 'default')}/jobs/{job_name}/runs/{{{{macros.OpenLineagePlugin.lineage_run_id(run_id, task)}}}}?api_key={os.environ.get('OPENLINEAGE_API_KEY')}"
+        dag=dag)
+```
+
 ## Development
 
 To install all dependencies for _local_ development:
